+++ packages/cli/run_examples.sh
@@ -2,7 +2,7 @@
 set -o errexit -o nounset -o pipefail
 command -v shellcheck >/dev/null && shellcheck "$0"
 
-if [ -n "${SIMAPP_ENABLED:-}" ]; then
+if [ -n "${SIMAPP42_ENABLED:-}" ]; then
   yarn node ./bin/cosmwasm-cli --init examples/cosmwasm.ts --code "process.exit(0)"
 fi
 if [ -n "${LAUNCHPAD_ENABLED:-}" ]; then
@@ -13,6 +13,6 @@ yarn node ./bin/cosmwasm-cli --init examples/generate_address.ts --code "process
 yarn node ./bin/cosmwasm-cli --init examples/local_faucet.ts --code "process.exit(0)"
 yarn node ./bin/cosmwasm-cli --init examples/mask.ts --code "process.exit(0)"
 yarn node ./bin/cosmwasm-cli --init examples/multisig_address.ts --code "process.exit(0)"
-if [ -n "${SIMAPP_ENABLED:-}" ]; then
+if [ -n "${SIMAPP42_ENABLED:-}" ]; then
   yarn node ./bin/cosmwasm-cli --init examples/stargate.ts --code "process.exit(0)"
 fi