+++ packages/stargate/src/queries/bank.spec.ts
@@ -4,6 +4,7 @@ import {
   nonExistentAddress,
   nonNegativeIntegerMatcher,
   pendingWithoutSimapp,
+  pendingWithoutSimapp42,
   simapp,
   unused,
 } from "../testutils.spec";