+++ packages/stargate/src/signingstargateclient.spec.ts
@@ -23,6 +23,7 @@ import {
   ModifyingDirectSecp256k1HdWallet,
   ModifyingSecp256k1HdWallet,
   pendingWithoutSimapp,
+  pendingWithoutSimapp42,
   simapp,
   validator,
 } from "./testutils.spec";
@@ -117,7 +118,7 @@ describe("SigningStargateClient", () => {
 
   describe("sendIbcTokens", () => {
     it("works with direct signing", async () => {
-      pendingWithoutSimapp();
+      pendingWithoutSimapp42();
       const wallet = await DirectSecp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const client = await SigningStargateClient.connectWithSigner(
         simapp.tendermintUrl,
@@ -166,7 +167,7 @@ describe("SigningStargateClient", () => {
     });
 
     it("works with Amino signing", async () => {
-      pendingWithoutSimapp();
+      pendingWithoutSimapp42();
       const wallet = await Secp256k1HdWallet.fromMnemonic(faucet.mnemonic);
       const client = await SigningStargateClient.connectWithSigner(
         simapp.tendermintUrl,